Based on a suspicion that ICARUS was having crashes related to iterator validity problems, I have updated MessageViewer to use shared_ptrs instead of iterator references in its data model. I have also reevaludated the mutexes, trying to get more granular mutexes to avoid over-locking.
I have also added a feature where if a user tries to apply a filter that
already is in use, they are simply taken to that tab instead of creating
multiple tabs with the same filter expression.

Only the active tab is actively updated, all other tabs simply have their
message lists updated and it is rendered when the tab becomes active.

The message limit is now applied per-tab; an old message will only be
discarded once it has fallen off all active filters. (It is still
removed from the message pool, so new filters will not see stale
messages.)